# express/src/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Express API",
    "version": "1.0.0",
    "description": ""
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": ["http"],
  "paths": {
    "/create_chat": {
      "post": {
        "tags": [
          "chats"
        ],
        "summary": "Create Chat",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "example": "chat name"
                },
                "ident": {
                  "example": "identifier"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "uuid": "hex num",
                  "name": "chat name",
                  "model": "{...model data}",
                  "model_ident": "identifier",
                  "project_uuid": "hex num",
                  "tokens": 0 
                }
              }
            }
          },
          "404": {
            "example": "No model loaded with this identifier",
            "description": "Not Found"
          }
        }
      }
    },
    "/chat/{uuid}": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "uuid",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "promt": {
                  "example": {
                    "role": "user",
                    "content": "string"
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/list_models": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/list_loaded_models": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/load_model": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "path": {
                  "example": "string"
                },
                "type": {
                  "example": "llm"
                },
                "contentLength": {
                  "example": 32000
                },
                "ident": {
                  "example": "identifier"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/unload_model": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "example": "llm"
                },
                "ident": {
                  "example": "identifier"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  }
}